README.md
Hello there!

Chi Hack Night wants to experiment with a new way of announcing of breakout groups and we could use your help! Here's the idea:

Currently, announcements for breakout groups are based off this semi-static list on the Chi Hack Night website. We'd like to try a more dynamic format using a list of GitHub issues like this one, which will allow breakout group leaders like yourself to more easily edit the written description of your projects so you can tailor it towards the skillsets you look to recruit from week to week.

Here's what you can do:

Add an issue to this repo for your project. If you can, follow the example in Issue #1, which has four sections broken out "About the group", "Who we're looking for", "Tools", and "Where we meet". You can also use the "labels" feature to tag your project as beginner-friendly or as in need of certain skillsets. We'll present this list of issues (hopefully in a more visually appealing format, Derek is working on that) next week and you can present concisely as normal.